[Order] Remove Adjustable trait
