+++ src/Sylius/Component/Order/Model/OrderItem.php
@@ -19,8 +19,6 @@ use Doctrine\Common\Collections\Collection;
  */
 class OrderItem implements OrderItemInterface
 {
-    use Adjustable;
-
     /**
      * @var mixed
      */
@@ -42,23 +40,11 @@ class OrderItem implements OrderItemInterface
     protected $unitPrice = 0;
 
     /**
-     * @var Collection|AdjustmentInterface[]
-     */
-    protected $adjustments;
-
-    /**
-     * @var int
-     */
-    protected $adjustmentsTotal = 0;
-
-    /**
      * @var int
      */
     protected $total = 0;
 
     /**
-     * Order item is immutable?
-     *
      * @var bool
      */
     protected $immutable = false;
@@ -69,8 +55,15 @@ class OrderItem implements OrderItemInterface
     protected $itemUnits;
 
     /**
-     * Constructor.
+     * @var Collection|AdjustmentInterface[]
      */
+    protected $adjustments;
+
+    /**
+     * @var int
+     */
+    protected $adjustmentsTotal = 0;
+
     public function __construct()
     {
         $this->adjustments = new ArrayCollection();
@@ -245,5 +238,101 @@ class OrderItem implements OrderItemInterface
     public function hasItemUnit(OrderItemUnitInterface $itemUnit)
     {
         return $this->itemUnits->contains($itemUnit);
+    }
+    /**
+     * {@inheritdoc}
+     */
+    public function getAdjustments($type = null)
+    {
+        if (null === $type) {
+            return $this->adjustments;
+        }
+
+        return $this->adjustments->filter(function (AdjustmentInterface $adjustment) use ($type) {
+            return $type === $adjustment->getType();
+        });
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function addAdjustment(AdjustmentInterface $adjustment)
+    {
+        if (!$this->hasAdjustment($adjustment)) {
+            $adjustment->setAdjustable($this);
+            $this->adjustments->add($adjustment);
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function removeAdjustment(AdjustmentInterface $adjustment)
+    {
+        if (!$adjustment->isLocked() && $this->hasAdjustment($adjustment)) {
+            $adjustment->setAdjustable(null);
+            $this->adjustments->removeElement($adjustment);
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function hasAdjustment(AdjustmentInterface $adjustment)
+    {
+        return $this->adjustments->contains($adjustment);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getAdjustmentsTotal($type = null)
+    {
+        if (null === $type) {
+            return $this->adjustmentsTotal;
+        }
+
+        $total = 0;
+        foreach ($this->getAdjustments($type) as $adjustment) {
+            $total += $adjustment->getAmount();
+        }
+
+        return $total;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function removeAdjustments($type)
+    {
+        foreach ($this->getAdjustments($type) as $adjustment) {
+            if ($adjustment->isLocked()) {
+                continue;
+            }
+
+            $this->removeAdjustment($adjustment);
+        }
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function clearAdjustments()
+    {
+        $this->adjustments->clear();
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function calculateAdjustmentsTotal()
+    {
+        $this->adjustmentsTotal = 0;
+
+        foreach ($this->adjustments as $adjustment) {
+            if (!$adjustment->isNeutral()) {
+                $this->adjustmentsTotal += $adjustment->getAmount();
+            }
+        }
     }
 }
